requirements.txt
aiohttp==3.11.14
aiohttp==3.12.12
beautifulsoup4==4.13.4
boilerpy3==1.0.7
crawl4ai==0.5.0.post6
fuzzywuzzy==0.18.0
jieba==0.42.1
langdetect==1.0.9
nltk==3.9.1
numpy==2.3.2
openai==1.98.0
openxlab==0.1.2
pandas==2.3.1
pyahocorasick==2.2.0
pydantic==2.11.7
python_dateutil==2.9.0.post0
Requests==2.32.4
schedule==1.2.2
selenium==4.34.2
torch==2.7.0
transformers==4.52.4
warcio==1.7.5
